requirements-test.txt
# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0

# Existing app dependencies (for testing)
streamlit>=1.28.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
Pillow>=10.0.0 